+++ R/ansi.R
@@ -1,5 +1,21 @@
 
-TRUE_COLORS <- as.integer(256^3)
+palette_idx <- function(id) {
+  ifelse(
+    id < 38,
+    id - (30 - 1),
+  ifelse(
+    id < 48,
+    -(id - (40 - 1)),
+  ifelse(
+    id < 98,
+    id - (90 - 9),
+    -(id - (100 - 9))
+  )))
+}
+
+palette_color <- function(x) {
+  c(x, palette = palette_idx(x[[1]]))
+}
 
 ansi_builtin_styles <- list(
   reset = list(0, c(0, 22, 23, 24, 27, 28, 29, 39, 49)),
@@ -11,42 +27,42 @@ ansi_builtin_styles <- list(
   hidden = list(8, 28),
   strikethrough = list(9, 29),
 
-  black = list(30, 39),
-  red = list(31, 39),
-  green = list(32, 39),
-  yellow = list(33, 39),
-  blue = list(34, 39),
-  magenta = list(35, 39),
-  cyan = list(36, 39),
-  white = list(37, 39),
+  black = palette_color(list(30, 39)),
+  red = palette_color(list(31, 39)),
+  green = palette_color(list(32, 39)),
+  yellow = palette_color(list(33, 39)),
+  blue = palette_color(list(34, 39)),
+  magenta = palette_color(list(35, 39)),
+  cyan = palette_color(list(36, 39)),
+  white = palette_color(list(37, 39)),
   silver = list(90, 39),
 
-  br_black = list(90, 39),
-  br_red = list(91, 39),
-  br_green = list(92, 39),
-  br_yellow = list(93, 39),
-  br_blue = list(94, 39),
-  br_magenta = list(95, 39),
-  br_cyan = list(96, 39),
-  br_white = list(97, 39),
+  br_black = palette_color(list(90, 39)),
+  br_red = palette_color(list(91, 39)),
+  br_green = palette_color(list(92, 39)),
+  br_yellow = palette_color(list(93, 39)),
+  br_blue = palette_color(list(94, 39)),
+  br_magenta = palette_color(list(95, 39)),
+  br_cyan = palette_color(list(96, 39)),
+  br_white = palette_color(list(97, 39)),
 
-  bg_black = list(40, 49),
-  bg_red = list(41, 49),
-  bg_green = list(42, 49),
-  bg_yellow = list(43, 49),
-  bg_blue = list(44, 49),
-  bg_magenta = list(45, 49),
-  bg_cyan = list(46, 49),
-  bg_white = list(47, 49),
+  bg_black = palette_color(list(40, 49)),
+  bg_red = palette_color(list(41, 49)),
+  bg_green = palette_color(list(42, 49)),
+  bg_yellow = palette_color(list(43, 49)),
+  bg_blue = palette_color(list(44, 49)),
+  bg_magenta = palette_color(list(45, 49)),
+  bg_cyan = palette_color(list(46, 49)),
+  bg_white = palette_color(list(47, 49)),
 
-  bg_br_black = list(100, 39),
-  bg_br_red = list(101, 39),
-  bg_br_green = list(102, 39),
-  bg_br_yellow = list(103, 39),
-  bg_br_blue = list(104, 39),
-  bg_br_magenta = list(105, 39),
-  bg_br_cyan = list(106, 39),
-  bg_br_white = list(107, 39),
+  bg_br_black = palette_color(list(100, 39)),
+  bg_br_red = palette_color(list(101, 39)),
+  bg_br_green = palette_color(list(102, 39)),
+  bg_br_yellow = palette_color(list(103, 39)),
+  bg_br_blue = palette_color(list(104, 39)),
+  bg_br_magenta = palette_color(list(105, 39)),
+  bg_br_cyan = palette_color(list(106, 39)),
+  bg_br_white = palette_color(list(107, 39)),
 
   # similar to reset, but only for a single property
   no_bold          = list(c(0,     23, 24, 27, 28, 29, 39, 49), 22),
@@ -97,19 +113,21 @@ ansi_style_str <- function(x) {
   paste0("\u001b[", x, "m", collapse = "")
 }
 
-create_ansi_style_tag <- function(name, open, close) {
+create_ansi_style_tag <- function(name, open, close, palette = NULL) {
   structure(
-    list(list(open = open, close = close)),
+    list(list(open = open, close = close, palette = palette)),
     names = name
   )
 }
 
 create_ansi_style_fun <- function(styles) {
   fun <- eval(substitute(function(...) {
-    mystyles <- .styles
     txt <- paste0(...)
-    if (num_ansi_colors() > 1) {
+    nc <- num_ansi_colors()
+    if (nc > 1) {
+      mystyles <- .styles
       for (st in rev(mystyles)) {
+        if (!is.null(st$palette)) st <- get_palette_color(st, nc)
         txt <- paste0(
           st$open,
           gsub(st$close, st$open, txt, fixed = TRUE),
@@ -129,7 +147,8 @@ create_ansi_style_fun <- function(styles) {
 create_ansi_style <- function(name, open = NULL, close = NULL) {
   open <- open %||% ansi_style_str(ansi_builtin_styles[[name]][[1]])
   close <- close %||% ansi_style_str(ansi_builtin_styles[[name]][[2]])
-  style <- create_ansi_style_tag(name, open, close)
+  palette <- ansi_builtin_styles[[name]]$palette
+  style <- create_ansi_style_tag(name, open, close, palette)
   create_ansi_style_fun(style)
 }
 
@@ -263,7 +282,7 @@ ansi_style_8_from_rgb <- function(rgb, bg) {
 
 ansi_style_from_rgb <- function(rgb, bg, num_colors, grey) {
   if (num_colors < 256) { return(ansi_style_8_from_rgb(rgb, bg)) }
-  if (num_colors < TRUE_COLORS || grey) return(ansi256(rgb, bg, grey))
+  if (num_colors < truecolor || grey) return(ansi256(rgb, bg, grey))
   return(ansitrue(rgb, bg))
 }